issue: 1438405 Parse vlan priority mapping file
This commit adds parsing the vlan mapping file using libnl,
with this information we can redirect the user setsockopt
to the right priority.

+++ src/vma/util/data_updater.h
@@ -58,7 +58,7 @@ public:
     virtual ~header_pcp_updater() {};
     virtual bool update_field(dst_entry &hdr);
 private:
-    uint8_t m_pcp;
+    uint32_t m_pcp;
 };
 
 class header_tos_updater: public data_updater {
